+++ CHANGELOG.md
@@ -5,6 +5,7 @@ Unreleased
 ----------
 
 - Added interface type='banner' into JSON schema.
+- Added feature "delete component".
 - Updated URLs to Custom Apps documentation.
 
 2.1.0 [2024-10-09]
@@ -14,6 +15,7 @@ Unreleased
 - Improved wizard of new module: Added questions for connection and alternative connection.
 - Added the in-editor validation of `makecomapp.json` project file.
 - Fixed the issue `"options.maxOutputLength" is out of range` occured during ZIP import.
+- Rapidly improved deployment of local app files. Only the changed data are being deployed now.
 - Added "Apply for all" into component mapping alignment process.
 - Added feature "remember origin and do not ask again".